+++ census/static/js/widget.geo.select.js
@@ -1,5 +1,4 @@
 var textmatchAPI = '/place-search/json/',
-    geocodingAPI = 'http://wards.code4sa.org/',
     resultTemplate = Handlebars.compile('<p class="result-name"><span class="result-type">{{geo_level}}</span>{{full_name}}</p>'),
     geoSelect = $('#geography-select, #geography-select-home');
 
@@ -20,36 +19,6 @@ var textMatchEngine = new Bloodhound({
 });
 textMatchEngine.initialize();
 
-var geoSelectEngine = new Bloodhound({
-    datumTokenizer: function(d) { return Bloodhound.tokenizers.whitespace(d.full_name); },
-    queryTokenizer: Bloodhound.tokenizers.whitespace,
-    limit: 20,
-    remote: {
-        url: geocodingAPI,
-        rateLimitWait: 600,
-        replace: function (url, query) {
-            return url += '?address=' + query + '&database=wards_2011';
-        },
-        filter: function(response) {
-            if (response.error !== undefined) return [];
-
-            return response.map(function(item) {
-                return {
-                    full_name: ['Ward ' + item['wards_no'] + ' (' + item['ward'] + ')',
-                                 item['municipality'], item['province']].join(', '),
-                    full_geoid: 'ward-' + item['ward'],
-                    geo_level: 'ward',
-                    geo_code: item['ward'],
-                };
-            });
-        },
-        ajax: {
-            dataType: 'jsonp',  // allow cross domain request
-        },
-    },
-});
-geoSelectEngine.initialize();
-
 function makeGeoSelectWidget(element, selected) {
     element.typeahead({
         autoselect: true,
@@ -61,15 +30,6 @@ function makeGeoSelectWidget(element, selected) {
         name: 'textmatch',
         displayKey: 'full_name',
         source: textMatchEngine.ttAdapter(),
-        limit: 20,
-        templates: {
-            suggestion: resultTemplate,
-        },
-    }, {
-        // get geocoded matches from Code4SA API
-        name: 'geocoded',
-        displayKey: 'full_name',
-        source: geoSelectEngine.ttAdapter(),
         limit: 20,
         templates: {
             suggestion: resultTemplate,