+++ CHANGES.rst
@@ -4,6 +4,8 @@ Wazimap Change Log
 0.2.3 (unreleased)
 ------------------
 
+* FEATURE: use Google place search to find places
+* FEATURE: new ``country_code`` config setting
 * FIX: redirect to slugged URLs when possible
 * BREAKING: Data tables now default to one table for all geo levels
 * Include deployment file examples in the ``deploy`` directory